containers/eve-api/requirements.txt
flask
Flask-PyMongo
mongoengine
flask_api
mongomock
pytz
pytest
pytest-cov
flask-mongoengine
python-dateutil
